Add a few instances of `pragma(inline, true)`
Based on the "AST" output from it, Clang at least appears to implicitly mark all of the functions inline in the C++ version, so it can't hurt to mark them explicitly here also (even though IMO these functions are unlikely to be inlineable very often). In any case, it's more likely to benefit DMD (which I don't think auto-inlines very much) than LDC (which presumably auto-inlines as much as Clang).